refactor app.py more, fix notification page
