+++ backend-flask/lib/cognito_jwt_token.py
@@ -6,6 +6,7 @@ from jose.utils import base64url_decode
 from functools import wraps, partial
 from flask import request, g
 import os
+from flask import current_app as app
 
 class FlaskAWSCognitoError(Exception):
     pass
@@ -137,7 +138,7 @@ def jwt_required(f=None, on_error=None):
             # unauthenticated request
             app.logger.debug(e)
             if on_error:
-                on_error(e)
+                return on_error(e)
             return {}, 401
         return f(*args, **kwargs)
     return decorated_function
